New utility method `Sniff::is_in_function_call()`
Allows for checking whether a token is (part of) a parameter passed to a specific (set of) function(s).

Optionally checks whether the function call is to a global function.
Optionally only checks the "deepest" function call.

Returns the stack pointer to the `T_STRING` token for the function call or `false`.