Sniff::has_nonce_check(): allow for comparing a variable before nonce check
This builds onto the similar changes made for the `ValidatedSanitizedInput` sniff in ...

This fixes false positives as reported in 1114 and 1506.

Note: it is not currently checked that the nonce check is done within the same conditional scope as the comparison. Just that it is done within the same _function_ scope.

Includes unit tests.

Fixes 1114
Fixes 1506