NonceVerification: improve handling of nested functions
The `Sniff::has_nonce_verification()` method would determine the scope within which to look for a nonce verification function based on the first function condition of the `$stackPtr` being examined.

This breaks when closures/functions are nested within other functions, as the code within those nested constructs will be ignored, while instead, that should be the code which should be examined.

By changing the start point of the scope within which the method looks for nonce verification functions to the deepest nested function, this use-case is handled correctly.

Note that PHPCS 3.5.0 is expected to contain a utility function which can largely replace the code which was now added. Once PHPCS 3.5.0 has been released and WPCS ups it's minimum version requirement, this change should be made.

Fixes 1694