Array sniffs: ignore short lists
While we should probably define rules for long/short list constructs, WPCS at this time, does not have a opinion on the formatting of these.

The `Arrays` sniffs, however, would all treat _short lists_ as if they were _short arrays_ and apply the array rules on them.

This PR fixes this by bowing out early if a short array is in actual fact a short list.

Includes unit tests in each of the sniffs in the `Arrays` category affected by this issue.

Fixes 1692