EscapeOutput: allow for typical pattern with `_deprecated_file()`
The first parameter passed to `_deprecated_file()` generally is `basename( __FILE__ )` based on the code currently in WP Core.
As the result of that function call is safe, I'm proposing making an exception for that particular code pattern.

Includes unit tests.

Note: the current code does not allow for comments in the first parameter. This would be rare encounter in these function calls anyway and allowance for it can be added later if needs be.