ArrayIndentation: bug fix - prevent fixer conflict
If an array item has something else on the same line before it, the indentation should be left alone.

This fixes a potential fixer conflict with the `DisallowInlineTabs` sniff which could occur when the sniffs are being run in a unconventional order.

Discovered when making a change to the Travis file and not reproducible locally, but consistently reproducible with the alternative running order of sniffs on select platforms, as per the Travis build on Xenial.