CronInterval: add the parameter name to allow for named parameters
PHPCSUtils 1.0.0-alpha4 supports named parameters in the `PassedParameters` class, but requires for the parameter name to be passed either as a string or an array (of multiple names in case of a renamed parameter).

The `CronInterval` sniff not passing the parameter name would cause the tests to fail - even when the tests do not contain a test for named parameters yet -, so let's fix the function call.

Note: this doesn't add proper support for named parameters to the sniff yet. That will be addressed later.