Move "deprecation" related utilities to dedicated `DeprecationHelper`
The "deprecation" related utilities are only used by a small set of sniffs, so are better placed in a dedicated class.

This commit moves the `is_function_deprecated()` method to a new `WordPressCS\WordPress\Helpers\DeprecationHelper` and starts using that class in the relevant sniffs.

**Note**:
It is expected for PHPCSUtils to have dedicated methods for the same at some point in the future. If/when those methods become available, it is recommended for the sniffs to start using the PHPCSUtils methods and for this class to be deprecated and removed in the next major release.

Related to 1465