PHP 8.0 | VariableHelper::is_comparison(): add support for variables being compared in a `match` expression
PHP 8.0 introduced `match` expressions, which do a (strict) comparison on the value in the condition against the options in the `match` body.
Ref: https://www.php.net/manual/en/control-structures.match.php

This commit updates the `is_comparison()` method to allow for variables in the "condition" part of a `match` expression to be considered part of a comparison.

Tested via the `ValidatedSanitizedInput` sniff.