NoSilencedErrors: rename property [3]
As this is a `protected` property, this is an internal change only (unless an external standard would be extending the sniff).

Also note that the new name is in line with the name for the same property in an upcoming replacement sniff in PHPCSExtra. This should make the switch over to the replacement sniff more straight-forward (if and when).