DB/PreparedSQL: use PHPCSUtils for checking string embeds
Switch out the use of the WPCS native `TextStringHelper::get_interpolated_variables()` method in favour of the significantly more comprehensive PHPCSUtils `TextStrings::getEmbeds()` method.

This improves the handling of embedded variables in double quoted strings (and consequently fixes some bugs).

For some samples of embedded variables/expressions, the sniff would previously correctly report on the embed, but would contain an incomplete/incorrect representation of the embed in the error message:
```php
$wpdb->query( "SELECT * FROM {$wpdb->bar()} WHERE post_title LIKE '{$title->bar()}';" );
//  106 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable $title at "SELECT * FROM {$wpdb->bar()} WHERE post_title LIKE '{$title->bar()}';"

$wpdb->query( "SELECT * FROM ${wpdb->{$baz}} WHERE post_title LIKE '${title->{$sub}}';" );
// 108 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable $sub at "SELECT * FROM ${wpdb->{$baz}} WHERE post_title LIKE '${title->{$sub}}';"
```

In other cases, the `$wpdb` variable was not correctly recognized as such (false positive):
```php
$wpdb->query( "SELECT * FROM ${wpdb->{$baz}} WHERE post_title LIKE '${title->{$sub}}';" );
//  108 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable $baz at "SELECT * FROM ${wpdb->{$baz}} WHERE post_title LIKE '${title->{$baz}}';"
```

For other embedded variables/expressions, the sniff would previously not throw an error (false negative).
```php
// No error, while an error should be thrown.
$wpdb->query( "SELECT * FROM ${wpdb->bar} WHERE post_title LIKE '${title->sub}';" ); // Bad x 1.
$wpdb->query( "SELECT * FROM ${wpdb->{${'a'}}} WHERE post_title LIKE '${title->{${'sub'}}}';" ); // Bad x 1.
```

With the implemented fix, all embeds will now result in the expected errors and the error messages will contain a correct representation of the embed:
```
 106 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable {$title->bar()} at "SELECT * FROM {$wpdb->bar()} WHERE post_title LIKE '{$title->bar()}';"
 107 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable ${title->bar} at "SELECT * FROM ${wpdb->bar} WHERE post_title LIKE '${title->bar}';"
 108 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable ${title->{$baz}} at "SELECT * FROM ${wpdb->{$baz}} WHERE post_title LIKE '${title->{$baz}}';"
 109 | ERROR | Use placeholders and $wpdb->prepare(); found interpolated variable ${title->{${'a'}}} at "SELECT * FROM ${wpdb->{${'a'}}} WHERE post_title LIKE '${title->{${'a'}}}';"
```

Note: the matching of `$wpdb` has been improved, but does not check whether a known table name is accessed or anything. This is no different than before, but may warrant further improvement in the future.