Arrays/ArrayKeySpacingRestrictions: improve handling of brackets for array assignments without key
As it was, this was handled correctly, though the error message was unclear as it referred to "array keys", while these assignments don't have an explicit key.

To fix this, I'm now special casing array assignments without an explicit key.

The original error message in that situation was: "Array keys must NOT be surrounded by spaces if they only contain a string or an integer."
The new error message will be: "There should be no space between the square brackets for an array assignment without an explicit key. Found: ..."

The new error message will also have a different, new error code: `SpacesBetweenBrackets`.
Note: the new error code could be considered a breaking change, though I doubt many people will notice this as this is not typically an error code which would be used in inline ignore annotations.

Includes unit tests.

Note: the behaviour for array assignments without an explicit key, but with a comment between the brackets, has not been changed. In that case, the original `NoSpacesAroundArrayKeys` error will still be thrown.

All the same, tests have been added to verify and safeguard this.