Core: move rules related to object instantiations from `Extra` to `Core`
> 1.  When instantiating a new instance of an object, parenthesis must always be used, even when not strictly necessary.
> 2. There should be no space between the name of the class being instantiated and the opening parenthesis.

Note: the spacing inside the parentheses when instantiating anonymous classes will not (yet) be checked by the `PEAR.Functions.FunctionCallSignature` sniff. A PR is open upstream to add support for this: squizlabs/PHP_CodeSniffer 3636
Once the upstream PR has been merged, WPCS will automatically inherit the support as this sniff is already included in WPCS.

The following will no longer be checked by WPCS:
> 3. Assigning the return value of an object instantiation by reference is not allowed.

This is a parse error since PHP 7.0, so should not be the concern of WPCS (and hasn't been added to the handbook for the same reason). See: https://3v4l.org/W2Qj6

---
As discussed in 1884, this means that JS support for checking object instantiations will be dropped, but as WPCS is generally not used for checking JS code, this should not be a problem.
---

Refs and related issues:
* https://make.wordpress.org/core/2020/03/20/updating-the-coding-standards-for-modern-php/ - Object instantiation section
* https://developer.wordpress.org/coding-standards/wordpress-coding-standards/php/#object-instantiation
* WordPress/wpcs-docs 109
* WordPress/WordPress-Coding-Standards 919
* WordPress/WordPress-Coding-Standards 1033
* WordPress/WordPress-Coding-Standards 1721
* squizlabs/PHP_CodeSniffer 3200
* squizlabs/PHP_CodeSniffer 3636 (not yet merged)
* PHPCSStandards/PHPCSExtra 76
* PHPCSStandards/PHPCSExtra 120

Fixes 1884
Fixes 2033