Core: add sniff to discourage using reserved keywords as param names
The sniff in PHPCSExtra throws a `warning`. This is in line with the wording now used in the WP Core CS handbook.

All the same, as warnings for the `src` directory are currently ignored in CI for WP Core, we could consider making this an `error` once all known instances have been fixed, which should be soon as the fixes are all lined up to be committed, and we _really_ don't want new ones being introduced (as renaming the parameter after introduction would be a BC-break).
For `Extra` we could then choose to make this less opinionated and revert it to a `warning`.

Refs:
* WordPress/wpcs-docs 120
* PHPCSStandards/PHPCSExtra 80

Related to WP Core tickets:
* https://core.trac.wordpress.org/ticket/51553
* https://core.trac.wordpress.org/ticket/55327
* https://core.trac.wordpress.org/ticket/55650
* https://core.trac.wordpress.org/ticket/56788