Core: ensure more keywords are lowercase
The `function` and `const` keywords when used in import `use` statements are tokenized as `T_STRING`, not their keyword token, which means that the `Generic.PHP.LowerCaseKeyword` overlooks these.

This adds an additional sniff, which specifically checks that the `function` and `const` keywords in import `use` statements are in lowercase.