Add sniff that will check that capabilities are used correctly.
The sniff will check if the functions that are accepting capabilities as the argument actually use capabilities and not roles.
It was rewritten to include helper functions from PHPCSUtils, cleanup code, and add additional tests.
The list of core functions that are using capabilities as a parameter was updated with changes up to WordPress 6.1.0,
as well as the list of capabilities.
The sniff is also compatible with PHP 8 (named arguments).

Co-authored-by: Juliette <663378+jrfnl@users.noreply.github.com>
Co-authored-by: Ulrich Pogson <grappler@users.noreply.github.com>
Co-authored-by: Kevin Haig <khacoder@users.noreply.github.com>
Co-authored-by: Gary Jones <github@garyjones.io>