MinimumWPVersionTrait: rename minimum WP version property
As discussed in 2113, this changes the name for both the CLI setting as well as the property to `minimum_wp_version` in all applicable places.

Note: documentation (`phpcs.xml.dist.sample` and the Wiki) is not being updated in this PR.
The wiki should be reviewed and updated just before release.
The update to the `phpcs.xml.dist.sample` file is included in a WIP PR to update all public facing documentation just before the release.

Closes 2113