GH Actions: tweak label management workflow
Let's try this without the quotes around the labels.
The example in the action runners' readme didn't contain quotes, I added them as the labels we use have a `:` in it, which has special meaning in YAML workflows, but clearly that's not working.

If this doesn't work either, I may need to report it as a bug to the action runner (or find another action runner for the same).