WP/EnqueuedResourceParameters: prevent false positives
The `'raw'` key in the parameter arrays returned from the `PassedParameters` class contains - as per the name - the _raw_ contents of the parameter.

Since PHPCSUtils 1.0.0-alpha4, the return array also contain a `'clean'` index, which contains the contents of the parameter cleaned of comments.

By switching to using that key, a `null` value accompanied by a comment will no longer throw the wrong error.

Includes unit test demonstrating the issue and safeguarding the fix.