PHP/NoSilencedErrors: implement PHPCSUtils methods
As the methods from `GetTokensAsString` work slightly differently from the PHPCS native `File::getTokensAsString()` method, this would result in a "one token longer" code snippet. The `-1` in the function call in line 225 has been added to prevent this change in behaviour without the need for people to update the `context_length` property value in their custom ruleset (which would be an unnecessary BC-break).

Functionally, using the `GetTokensAsString::compact()` method will result in the code snippet in the error message being "cleaner" as those will no longer contain comments found in the code. The PHPCSUtils method will also "compact" the whitespace, so no need for the `str_replace()` call anymore.

No additional tests added, the existing tests (line 58) already cover this.

Note: it is still the intention for a similar sniff to become available via PHPCSExtra, but that will be in a later release.