NamingConventions/ValidHookName::transform_complex_string(): simplify with PHPCSUtils
Simplify the `transform_complex_string()` method by using the PHPCSUtils utilities to split a text string containing interpolated variables or expressions, instead of using custom logic to do this in the sniff.

Functionally this has no impact. The "old" code already handled complex expressions correctly.
Performance-wise, this could (should) have a small, positive impact as PHPCSUtils caches the results of the splitting of the text string and using the Utils functions allows for simplifying the logic in the function itself, which is called three times for each double quoted teẋt string.

Includes a few extra tests with some more complex variables/expressions.