WP/AlternativeFunctions: minor code tweak / readfile/fopen/file_put_contents()
The `'raw'` key in the parameter arrays returned from the `PassedParameters` class contains - as per the name - the _raw_ contents of the parameter.

Since PHPCSUtils 1.0.0-alpha4, the return array also contain a `'clean'` index, which contains the contents of the parameter cleaned of comments.

While this doesn't make a difference functionally (no false positives/negatives), it may improve performance a tiny bit as there is less text for the `is_local_data_stream()` method to handle.