WP/GlobalVariablesOverride: add support for PHP 7.4+ arrow functions
PHP 7.4 introduced arrow functions.

For the purposes of this sniff, default value assignments to parameters declared in an arrow function should be ignored to prevent false positives.

I've also reviewed the other places where _special handling_ is in place for functions/closures and have concluded that this does not apply to arrow functions.

I have updated line 333 to use the PHPCSUtils `Collections::functionDeclarationTokens()` token array for convenience, but it won't have any effect on arrow functions as `global` statements are not supported in arrow functions (well, that is, I haven't been able to set up any code which didn't result in a parse error for that. May need revisiting if someone more creative than me comes up with a way to do it...)

Includes unit test for the parameter default value part.