NamingConventions/PrefixAllGlobals: (mostly) ignore PHP 7.4+ arrow functions
PHP 7.4 introduced arrow functions.

For the purposes of this sniff:
* Parameters declared by arrow functions (with a default value assignment) do not need to be prefixed.
* New variables declared within an arrow function are local to the arrow function, so can be ignored.
* A `global` statement is not allowed within an arrow function.

And as an arrow function can only contain one statement, the chances of anything else within the arrow function needing to be flagged are slim, except when an arrow function declares a closure.
Note: nesting a named function declaration in an arrow function is not allowed (parse error).

PHPCS adds the `parenthesis_*` keys correctly and adds the arrow function parentheses to the `nested_parenthesis` array, so the function parameters can be handled same like parameters for closures/named functions.

However, while PHPCS does add the `scope_*` keys to the `T_FN` function, it does **not** add the `T_FN` to the `conditions` array for tokens between the `scope_opener` and the `scope_closer`.
The reason for this, is that the `scope_closer` of arrow functions can overlap with other `scope_closer`s, like in nested arrow function, which would make the `conditions` array unusable.

With the above in mind, I'm proposing to skip over the contents of arrow functions and ignore everything in them completely, with the one exception, that if an arrow function declares a closure, the closure will still be examined.

While this could potentially lead to some false negatives, I do believe this is the best we can do for now.

Includes unit tests.