NamingConventions/PrefixAllGlobals: add support for PHP 8.1+ enums
PHP 8.1 introduced enums as a new OO structure.

Unless in a (prefixed) namespace, enum names should also be prefixed.

The `Tokens::$ooScopeTokens` array used in the `register()` method already contains the `T_ENUM` token, so the sniff is already listening for it.

This commit now adjusts the sniff to actually handle enums as well.

Includes additional unit tests.

Note: enums cannot `extend`, so cannot ever be regarded as "test classes" for the purposes of this sniff.