NamingConventions/PrefixAllGlobals: minor bug fix
The regex as-was allowed for the `x7f` (= DEL) character to be part of a name.

The origin of this bug was the PHP manual, where the regexes were originally incorrect. This was fixed in https://github.com/php/doc-en/commit/6187696fb48916b4585a8a837bf725845f7cf3c7

No unit tests included as, to be honest, I haven't got a clue how to add a test with the DEL character ;-)

Includes adding a note about why this condition is not using PHPCSUtils.