DB/PreparedSQL: use predefined token collections [2]
Leverage predefined token collections in PHPCS and PHPCSUtils to fill the list of `$ignored_tokens`.

This commit start using the PHPCS native `Tokens::$castTokens` collection.

The `Tokens::$castTokens` collection contains _all_ cast tokens, not just the int/float/boolean cast tokens. However, as a cast token, in and of itself, is not problematic, I believe it is safe to include these.

What I mean by this, is that the sniff shouldn't trigger on seeing `(object)` in an SQL query. The sniff should trigger on the `$var`, which is being cast to an object, as an object cast is not a "safe" cast for use in SQL queries.

Having those extra cast tokens in the "ignore" list, therefore doesn't make the sniff any less effective as the sniff has a separate check for whether or not a type cast is considered "safe".

Includes a test demonstrating the issue.