CodeAnalysis/AssignmentInCondition: defer to upstream sniff
Upstream PR squizlabs/PHP_CodeSniffer 1594 pulled largely the same sniff to PHPCS itself and was merged in PHPCS 3.1.0.

The only part of the sniff I held back at the time, was the "is this an assignment in a ternary condition" part of the sniff as that was (and still is) more prone to false positives/negatives due to the difficulty in determining the start and end of the condition.

As the minimum supported PHPCS version for WPCS is now well beyond PHPCS 3.1.0, I'd like to propose to defer to the upstream sniff for the control structure condition checking and to only keep the code related to checking for assignments in ternary conditions in WPCS.

To prevent confusion between the two sniffs (the upstream one and this one) and to make it clear which is checking what, I also propose to rename the sniff to `WordPress.CodeAnalysis.AssignmentInTernaryCondition`.

Includes bringing the WPCS native PHPCS ruleset inline with this change.