WP/I18n: bug fix - fixer for `UnorderedPlaceholders*` could mangle text string
As things were, the fixer for the `UnorderedPlaceholders*` error, would replace the first non-empty token in the parameter. While this is fine in the majority of cases, this would lead to a mangled text string in the `fixed` file for multi-line text strings.

Fixed now.

Includes unit test safeguarding the fix.